+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/AddCacheClearerPass.php
@@ -11,31 +11,17 @@
 
 namespace Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler;
 
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\Reference;
+@trigger_error(sprintf('The %s class is deprecated since version 3.4 and will be removed in 4.0. Use Symfony\Component\HttpKernel\DependencyInjection\AddCacheClearerPass instead.', AddCacheClearerPass::class), E_USER_DEPRECATED);
+
+use Symfony\Component\HttpKernel\DependencyInjection\AddCacheClearerPass as BaseAddCacheClearerPass;
 
 /**
  * Registers the cache clearers.
  *
+ * @deprecated since version 3.4, to be removed in 4.0. Use {@link BaseAddCacheClearerPass instead}.
+ *
  * @author Dustin Dobervich <ddobervich@gmail.com>
  */
-class AddCacheClearerPass implements CompilerPassInterface
+class AddCacheClearerPass extends BaseAddCacheClearerPass
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function process(ContainerBuilder $container)
-    {
-        if (!$container->hasDefinition('cache_clearer')) {
-            return;
-        }
-
-        $clearers = array();
-        foreach ($container->findTaggedServiceIds('kernel.cache_clearer', true) as $id => $attributes) {
-            $clearers[] = new Reference($id);
-        }
-
-        $container->getDefinition('cache_clearer')->replaceArgument(0, $clearers);
-    }
 }